Quash various warnings
* include/ach/klinux_generic.h: signed type of ach_get options in
  ioctl

* src/ach_posix.c:
  (charfile_unlink): return enum ach_status and use use strncpy()
  (ach_get): const-correct pointer to timspec
  (achk_put): pass buffer as void* to achk_put()

* include/ach/impl_generic.h: (ach_create_len): return size_t

* src/achtool.c: fix return types

* src/achcat.c: don't error check with assert()

* src/ach-bench.c: don't error check with assert()

* src/test/achtest.c: remove unused variables

* src/achlog.c: remove unused variables

* src/achd/transport.c: remove unused variables

* src/achd/achd.c: replace assert(0) with abort()

* src/test/canceltest.c: quash unused argument warning

* src/test/transfertest.c: quash unused argument warning

+++ include/ach/impl_generic.h
@@ -114,7 +114,7 @@ static void free_index(ach_header_t *shm, size_t i )
 }
 
 
-static int ach_create_len( size_t frame_cnt, size_t frame_size )
+static size_t ach_create_len( size_t frame_cnt, size_t frame_size )
 {
     return sizeof(struct ach_header) +
         frame_cnt * sizeof(ach_index_t) +
